+++ app/models/user.rb
@@ -48,6 +48,9 @@ class User < ApplicationRecord
                                              message: "can only contain letters and spaces" }
 
   validates :email, presence: true, format: /\A[^@,\s]+@[^@,\s]+\.[^@,\s]+\z/
+  validates :name, length: { minimum: 1 }
+  validates :name, length: { maximum: 500 }
+  validate :clean_name
 
   scope :subscribed, -> { where(subscribed: true) }
 
@@ -113,5 +116,25 @@ class User < ApplicationRecord
 
     def purge_profile_picture
       profile_picture.purge if profile_picture.attached?
+    end
+
+    def clean_name
+      matchlists = %i[profanity hate violence]
+      language_filters = matchlists.map { |list| LanguageFilter::Filter.new(matchlist: list) }
+
+      description_matches = check_language_filters(language_filters, name)
+
+      return nil if description_matches.empty?
+
+      errors.add(
+        :description,
+        "contains inappropriate language: #{description_matches.join(', ')}"
+      )
+    end
+
+    def check_language_filters(filters, text)
+      filters.flat_map do |filter|
+        filter.matched(text) if filter.match?(text)
+      end.compact.uniq
     end
 end
